refactor how repo-level vendor directory is handled
Mounting the repo-level vendor directory inside the image (either
at the global location, or by overriding /app/vendor [where `app`
is really the current directory on the host]) introduces several
problems, namely, that post-install hooks will not be executed at the
root level, but rather in the current host directory. Additionally, the
composer.json and composer.lock files in the repository root will not be
used.

This changeset refactors what happens when a vendor directory is found
in the repository's root. If we find a vendor directory in the
repository's root, we assume that we are inside of a monolithic
repository, and that the repo-root really where `composer` should be
executed -- so we switch into it, and out back to our previous directory
when the command has completed execution.

closes gh-12